+++ org.qiweb/org.qiweb.api/src/main/java/org/qiweb/filters/ContentSecurityPolicy.java
@@ -18,6 +18,7 @@ package org.qiweb.filters;
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Inherited;
+import java.lang.annotation.Repeatable;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
@@ -71,6 +72,7 @@ import static org.qiweb.api.util.Strings.hasTextOrNull;
 @Retention( RetentionPolicy.RUNTIME )
 @Inherited
 @Documented
+@Repeatable( ContentSecurityPolicy.Repeat.class )
 public @interface ContentSecurityPolicy
 {
     String policy() default EMPTY;
@@ -159,5 +161,14 @@ public @interface ContentSecurityPolicy
             }
             return outcomes().noContent().build();
         }
+    }
+
+    @Retention( RetentionPolicy.RUNTIME )
+    @Target( { ElementType.METHOD, ElementType.TYPE } )
+    @Inherited
+    @Documented
+    public static @interface Repeat
+    {
+        ContentSecurityPolicy[] value();
     }
 }