+++ org.qiweb/org.qiweb.runtime/src/main/java/org/qiweb/runtime/filters/FilterChainFactory.java
@@ -16,6 +16,9 @@
 package org.qiweb.runtime.filters;
 
 import java.lang.annotation.Annotation;
+import java.lang.annotation.Repeatable;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.LinkedHashSet;
 import java.util.List;
@@ -28,6 +31,7 @@ import org.qiweb.api.filters.Filter;
 import org.qiweb.api.filters.FilterChain;
 import org.qiweb.api.filters.FilterWith;
 import org.qiweb.api.util.Couple;
+import org.qiweb.runtime.exceptions.QiWebRuntimeException;
 import org.qiweb.runtime.filters.FilterChainInstance.FilterChainControllerTail;
 
 /**
@@ -70,14 +74,16 @@ public class FilterChainFactory
         {
             if( FilterWith.class.equals( annotation.annotationType() ) )
             {
+                // Direct @FilterWith usage
                 FilterWith filterWith = (FilterWith) annotation;
                 for( Class<? extends Filter> filterClass : filterWith.value() )
                 {
                     filters.add( Couple.leftOnly( filterClass ) );
                 }
             }
-            else
+            else if( annotation.annotationType().getAnnotation( FilterWith.class ) != null )
             {
+                // Annotation annotated with @FilterWith usage
                 FilterWith filterWith = annotation.annotationType().getAnnotation( FilterWith.class );
                 if( filterWith != null )
                 {
@@ -85,6 +91,48 @@ public class FilterChainFactory
                     {
                         filters.add( Couple.of( filterClass, annotation ) );
                     }
+                }
+            }
+            else
+            {
+                // Repeatable annotation usage
+                Method[] methods = annotation.annotationType().getDeclaredMethods();
+                if( methods.length != 1
+                    || !"value".equals( methods[0].getName() )
+                    || !methods[0].getReturnType().isArray() )
+                {
+                    continue;
+                }
+                try
+                {
+                    Object[] array = (Object[]) methods[0].invoke( annotation );
+                    List<Annotation> repeatedAnnotations = new ArrayList<>();
+                    for( Object inner : array )
+                    {
+                        try
+                        {
+                            Annotation innerAnnotation = (Annotation) inner;
+                            if( innerAnnotation.annotationType().getAnnotation( Repeatable.class ) != null
+                                && annotation.annotationType().isAssignableFrom(
+                                    innerAnnotation.annotationType().getAnnotation( Repeatable.class ).value()
+                                ) )
+                            {
+                                repeatedAnnotations.add( innerAnnotation );
+                            }
+                        }
+                        catch( ClassCastException ex )
+                        {
+                        }
+                    }
+                    filters.addAll(
+                        findFilters(
+                            repeatedAnnotations.toArray( new Annotation[ repeatedAnnotations.size() ] )
+                        )
+                    );
+                }
+                catch( IllegalAccessException | IllegalArgumentException | InvocationTargetException ex )
+                {
+                    throw new QiWebRuntimeException( ex );
                 }
             }
         }